Welcome to my Github README.md
I'm an Infrastructure and Network Security Administrator strategically, aiming for senior-level expertise in IT risk, compliance, and cloud security. With a background in firewall and network administration, I’m currently helping my organization transition to a hybrid cloud model. My focus is on mastering risk and compliance frameworks tailored for financial and banking sectors, and I am preparing for key certifications, including CRISC, CISSP, and Azure AZ-500, to support this path. Driven by a commitment to secure, compliant solutions, I’m eager to collaborate on projects that bring value to these crucial fields.